services/google_sync/requirements.txt
notion-client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib